refactor(agentic-ai): move GatewayToolDefinitionResolver to schema package
